Hide fsevents representation
