+++ crates/extension/src/extension_store.rs
@@ -296,10 +296,10 @@ impl ExtensionStore {
             let reload_tx = this.reload_tx.clone();
             let installed_dir = this.installed_dir.clone();
             async move {
-                let mut events = fs.watch(&installed_dir, FS_WATCH_LATENCY).await;
-                while let Some(events) = events.next().await {
-                    for event in events {
-                        let Ok(event_path) = event.path.strip_prefix(&installed_dir) else {
+                let mut paths = fs.watch(&installed_dir, FS_WATCH_LATENCY).await;
+                while let Some(paths) = paths.next().await {
+                    for path in paths {
+                        let Ok(event_path) = path.strip_prefix(&installed_dir) else {
                             continue;
                         };
