+++ crates/fsevent/Cargo.toml
@@ -15,8 +15,6 @@ doctest = false
 [dependencies]
 bitflags.workspace = true
 parking_lot.workspace = true
-
-[target.'cfg(target_os = "macos")'.dependencies]
 fsevent-sys = "3.0.2"
 
 [dev-dependencies]
